pyproject.toml
[tool.poetry]
name = "repl_nix_alfafood"
version = "0.1.0"
description = ""
authors = ["Replit user <<>>"]

[tool.poetry.dependencies]
python = "^3.10"
django-rest-knox = "^4.2.0"
djangorestframework = "^3.14.0"
Pillow = "^9.5.0"
Django = "^4.2"
django-admin-rangefilter = "^0.10.0"
django-filter = "^23.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
